requirements.txt
flask==3.0.3
python-dotenv==1.1.1
crewai==1.11.1
langchain-core==0.3.63
langchain-community==0.3.20
langchain-huggingface==0.1.2
langchain==0.3.21
langgraph==0.2.50
faiss-cpu==1.9.0
sentence-transformers==3.3.1
httpx==0.28.1
litellm
